requirements.txt
torch
scikit-learn
torchvision 
Pillow 
streamlit 
pyserial
onnx
onnxruntime 
tensorboard